Task complete: work out if using a singleton for Db could be problematic
Apart from the `SQLiteDatabase` instance `Db.db` and the `Context` instance used
when creating `Db`, `Db` is stateless.

According to https://developer.android.com/reference/android/database/sqlite/SQLiteOpenHelper.html#getWritableDatabase%28%29
the  underlying `SQLiteDatabase` instance is cached and shared anyway.

The significance of the `Context` used when creating the initial instance of Db
is that the database relates to the package which the Context is from.  We are
always using this database from the same package, so this should not prevent any
issues.